+++ packages/polythene-react/scripts/standalone.js
@@ -1,7 +1,7 @@
 export * from "../index";
 export { subscribe, unsubscribe } from "polythene-core";
 import { addTypography } from "polythene-style";
-import { addLayoutStyles } from "polythene-utilities";
+import { addLayoutStyles } from "polythene-core-css";
 import { addFastClick } from "polythene-fastclick";
 
 addTypography();