+++ packages/polythene-utilities/src/index.js
@@ -1,5 +1,4 @@
 export { addWebFont } from "./webfont-loader";
 export { easing } from "./easing";
-export { layoutStyles, addLayoutStyles } from "./layout-styles";
 export { scrollTo } from "./scroll-to";
 export { Timer } from "./timer";